requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.23.2
pydantic==2.9.2
python-multipart==0.0.9
aiofiles==24.1.0
streamlit==1.35.0
requests==2.31.0
python-dotenv==1.0.1

# NEW Google Gemini SDK (must be latest)
google-generativeai>=0.5.3

# For JSON5 reading (your app uses it)
json5==0.9.14
